separate functions in sync.ts file
